// src/descriptions/DecisionDescription.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface DecisionDescription {
  id: string;
  label?: string;
  /**
   * Expr returning collection to choose from
   */
  from?: string;
  min?: number;
  max?: number;
  constrains?: ExprDescription[];
  effects?: EffectDescription[];
  dependsOn?: string[];
  /**
   * Die id for roll-dice decisions
   */
  dice?: string;
  [k: string]: unknown;
}
export interface ExprDescription {
  expr: string;
  explain?: string;
  [k: string]: unknown;
}
export interface EffectDescription {
  expr: string;
  chance?: number;
  if?: ExprDescription;
  [k: string]: unknown;
}